lots of renaming classes
fixed bug where operators were consistently misidentified as the operator whose symbol is the shortest substring of the correct operator
