--- FOLQuantifier.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace FYP_FC_Evaluator___UI
-{
-	enum FOLQuantifier
-	{
-		A,
-		E
-	}
-}
