+++ FYP FC Evaluator + UI.csproj
@@ -12,6 +12,8 @@
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
     <Deterministic>true</Deterministic>
+    <NuGetPackageImportStamp>
+    </NuGetPackageImportStamp>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
     <PlatformTarget>AnyCPU</PlatformTarget>
@@ -33,6 +35,15 @@
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="StyleCop, Version=6.2.0.0, Culture=neutral, PublicKeyToken=f904653c63bc2738, processorArchitecture=MSIL">
+      <HintPath>packages\StyleCop.6.2.0\lib\net40\StyleCop.dll</HintPath>
+    </Reference>
+    <Reference Include="StyleCop.CSharp, Version=6.2.0.0, Culture=neutral, PublicKeyToken=f904653c63bc2738, processorArchitecture=MSIL">
+      <HintPath>packages\StyleCop.CSharp.6.2.0\lib\net40\StyleCop.CSharp.dll</HintPath>
+    </Reference>
+    <Reference Include="StyleCop.CSharp.Rules, Version=6.2.0.0, Culture=neutral, PublicKeyToken=f904653c63bc2738, processorArchitecture=MSIL">
+      <HintPath>packages\StyleCop.CSharp.Rules.6.2.0\lib\net40\StyleCop.CSharp.Rules.dll</HintPath>
+    </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
     <Reference Include="System.Xml.Linq" />
@@ -47,11 +58,11 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Assignment.cs" />
-    <Compile Include="BottomUpFOLEvaluator.cs" />
+    <Compile Include="FOLBottomUpEvaluator.cs" />
     <Compile Include="FOLExpression.cs" />
-    <Compile Include="FOLOperator.cs" />
+    <Compile Include="Operator.cs" />
     <Compile Include="FOLParser.cs" />
-    <Compile Include="FOLQuantifier.cs" />
+    <Compile Include="Quantifier.cs" />
     <Compile Include="FOLSearcher.cs" />
     <Compile Include="FOLSearcherForm.cs">
       <SubType>Form</SubType>
@@ -59,14 +70,14 @@
     <Compile Include="FOLSearcherForm.Designer.cs">
       <DependentUpon>FOLSearcherForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="ParseTreeFOL_O.cs" />
-    <Compile Include="ParseTreeFOL_Q.cs" />
+    <Compile Include="ParseTreeOperatorNode.cs" />
+    <Compile Include="ParseTreeQuantifierNode.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="StringLogicComparator.cs" />
     <Compile Include="Term.cs" />
     <Compile Include="TermEvaluator.cs" />
-    <Compile Include="TopDownFOLEvaluator.cs" />
+    <Compile Include="FOLTopDownEvaluator.cs" />
     <EmbeddedResource Include="FOLSearcherForm.resx">
       <DependentUpon>FOLSearcherForm.cs</DependentUpon>
       <SubType>Designer</SubType>
@@ -80,6 +91,7 @@
       <AutoGen>True</AutoGen>
       <DependentUpon>Resources.resx</DependentUpon>
     </Compile>
+    <None Include="packages.config" />
     <None Include="Properties\Settings.settings">
       <Generator>SettingsSingleFileGenerator</Generator>
       <LastGenOutput>Settings.Designer.cs</LastGenOutput>
@@ -89,9 +101,18 @@
       <DependentUpon>Settings.settings</DependentUpon>
       <DesignTimeSharedInput>True</DesignTimeSharedInput>
     </Compile>
+    <None Include="Settings.StyleCop" />
+    <None Include="StyleCop.CSharp.Rules.targets" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
+  <Import Project="packages\StyleCop.CSharp.Rules.6.2.0\build\StyleCop.CSharp.Rules.targets" Condition="Exists('packages\StyleCop.CSharp.Rules.6.2.0\build\StyleCop.CSharp.Rules.targets')" />
+  <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
+    <PropertyGroup>
+      <ErrorText>This project references NuGet package(s) that are missing on this computer. Use NuGet Package Restore to download them.  For more information, see http://go.microsoft.com/fwlink/?LinkID=322105. The missing file is {0}.</ErrorText>
+    </PropertyGroup>
+    <Error Condition="!Exists('packages\StyleCop.CSharp.Rules.6.2.0\build\StyleCop.CSharp.Rules.targets')" Text="$([System.String]::Format('$(ErrorText)', 'packages\StyleCop.CSharp.Rules.6.2.0\build\StyleCop.CSharp.Rules.targets'))" />
+  </Target>
 </Project>
