+++ Operator.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace FYP_FC_Evaluator
+{
+	enum Operator
+	{
+		Disjunction,
+		Conjunction
+	}
+}
