+++ packages.config
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="StyleCop" version="6.2.0" targetFramework="net472" />
+  <package id="StyleCop.CSharp" version="6.2.0" targetFramework="net472" />
+  <package id="StyleCop.CSharp.Rules" version="6.2.0" targetFramework="net472" />
+</packages>
